+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `from_observation` class methods to local and global stat classes
+- `==` operator for `WeightsMatrix`
+
 ## [1.0.3] - 2020-05-22
 
 ### Added